+++ gobblin-service/src/main/java/org/apache/gobblin/service/modules/orchestration/proc/DagProcUtils.java
@@ -60,30 +60,22 @@ public class DagProcUtils {
   /**
    * If there is a single job to run next, it runs it. If there are multiple jobs to run, it creates a
    * {@link org.apache.gobblin.service.modules.orchestration.DagActionStore.DagActionType#REEVALUATE} dag action for
-   * each of them and those jobs will be launched in respective Reevaluate dag proc.
+   * each of them and those jobs will be launched in respective {@link ReevaluateDagProc}.
    */
   public static void submitNextNodes(DagManagementStateStore dagManagementStateStore, Dag<JobExecutionPlan> dag,
       DagManager.DagId dagId) throws IOException {
     Set<Dag.DagNode<JobExecutionPlan>> nextNodes = DagManagerUtils.getNext(dag);
 
-    if (nextNodes.size() > 1) {
-      handleMultipleJobs(dagManagementStateStore, nextNodes);
-      return;
-    }
-
-    //Submit jobs from the dag ready for execution.
-    for (Dag.DagNode<JobExecutionPlan> dagNode : nextNodes) {
+    if (nextNodes.size() == 1) {
+      Dag.DagNode<JobExecutionPlan> dagNode = nextNodes.iterator().next();
       DagProcUtils.submitJobToExecutor(dagManagementStateStore, dagNode, dagId);
       log.info("Submitted job {} for dagId {}", DagManagerUtils.getJobName(dagNode), dagId);
-    }
-  }
-
-  private static void handleMultipleJobs(DagManagementStateStore dagManagementStateStore, Set<Dag.DagNode<JobExecutionPlan>> nextNodes)
-      throws IOException {
-    for (Dag.DagNode<JobExecutionPlan> dagNode : nextNodes) {
-      JobExecutionPlan jobExecutionPlan = dagNode.getValue();
-      dagManagementStateStore.addJobDagAction(jobExecutionPlan.getFlowGroup(), jobExecutionPlan.getFlowName(),
-          String.valueOf(jobExecutionPlan.getFlowExecutionId()), jobExecutionPlan.getJobName(), DagActionStore.DagActionType.REEVALUATE);
+    } else {
+      for (Dag.DagNode<JobExecutionPlan> dagNode : nextNodes) {
+        JobExecutionPlan jobExecutionPlan = dagNode.getValue();
+        dagManagementStateStore.addJobDagAction(jobExecutionPlan.getFlowGroup(), jobExecutionPlan.getFlowName(),
+            String.valueOf(jobExecutionPlan.getFlowExecutionId()), jobExecutionPlan.getJobName(), DagActionStore.DagActionType.REEVALUATE);
+      }
     }
   }
 
@@ -154,6 +146,8 @@ public class DagProcUtils {
       }
       throw new RuntimeException(e);
     }
+
+    log.info("Submitted job {} for dagId {}", DagManagerUtils.getJobName(dagNode), dagId);
   }
 
   public static void cancelDagNode(Dag.DagNode<JobExecutionPlan> dagNodeToCancel, DagManagementStateStore dagManagementStateStore) throws IOException {
