+++ gobblin-service/src/main/java/org/apache/gobblin/service/modules/spec/JobExecutionPlan.java
@@ -68,6 +68,10 @@ public class JobExecutionPlan {
 
   private final JobSpec jobSpec;
   private final SpecExecutor specExecutor;
+  // the field is only read in one place in ResumeDagProc, in other places, a dag node's status is queried
+  // through API DagManagementStateStore#getDagNodeWithJobStatus which retrieves status directly from JobStatusRetriever
+  // todo - we should either keep this field updated as soon as a status is received so it can be used everywhere
+  //  or consider removing it completely
   private ExecutionStatus executionStatus = ExecutionStatus.PENDING;
   private final int maxAttempts;
   private int currentGeneration = 1;
